+++ GeosCore/fullchem_mod.F90
@@ -179,8 +179,7 @@ CONTAINS
     INTEGER                :: P,          MONTH,     YEAR,     Day
     INTEGER                :: IERR,       S,         Thread
     INTEGER                :: errorCount, previous_units
-    REAL(fp)               :: SO4_FRAC,   T,         TIN
-    REAL(fp)               :: TOUT,       SR,        LWC
+    REAL(fp)               :: SO4_FRAC,   SR,        LWC
     REAL(dp)               :: KPPH_before_integrate
     ! Strings
     CHARACTER(LEN=255)     :: errMsg,     thisLoc
@@ -236,9 +235,10 @@ CONTAINS
 !
     ! Defines the slot in which the H-value from the KPP integrator is stored
     ! This should be the same as the value of Nhnew in gckpp_Integrator.F90
-    ! (assuming Rosenbrock solver).  Define this locally in order to break
-    ! a compile-time dependency.  -- Bob Yantosca (05 May 2022)
+    ! Define this locally in order to break a compile-time dependency.
+    !    -- Bob Yantosca (05 May 2022)
     INTEGER,     PARAMETER :: Nhnew = 3
+
     ! Add Nhexit, the last timestep length -- Obin Sturm (30 April 2024)
     INTEGER,     PARAMETER :: Nhexit = 2
 
@@ -331,6 +331,7 @@ CONTAINS
 
     !========================================================================
     ! Zero out certain species
+    ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
     !========================================================================
     DO N = 1, State_Chm%nSpecies
 
@@ -426,6 +427,7 @@ CONTAINS
 
     !=======================================================================
     ! Archive concentrations before chemistry
+    ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
     !=======================================================================
     IF ( State_Diag%Archive_ConcBeforeChem ) THEN
        ! Point to mapping obj specific to ConcBeforeChem diagnostic collection
@@ -477,31 +479,12 @@ CONTAINS
     ! probably safe to define them here outside the OpenMP loop.
     ! (bmy, 3/28/16)
     !
-    ! The ICNTRL vector specifies options for the solver.  We can
-    ! define ICNTRL outside of the "SOLVE CHEMISTRY" parallel loop
-    ! below because it is passed to KPP as INTENT(IN).
-    ! (bmy, 3/28/16)
-    !
-    ! ICNTRL now needs to be updated within the TimeLoop for the AR solver
-    ! (hplin, 4/13/22)
+    ! ICNTRL and RCNTRL now need to be updated within the TimeLoop for
+    ! the Rosenbrock_AutoReduce solver (hplin, 4/13/22)
     !========================================================================
-
-    !%%%%% TIMESTEPS %%%%%
-
-    ! mje Set up conditions for the integration
-    ! mje chemical timestep and convert it to seconds.
-    DT        = GET_TS_CHEM() ! [s]
-    T         = 0d0
-    TIN       = T
-    TOUT      = T + DT
-
-    !%%%%% CONVERGENCE CRITERIA %%%%%
-
-    ! Absolute tolerance
-    ATOL      = State_Chm%KPP_AbsTol
-
-    ! Relative tolerance
-    RTOL      = State_Chm%KPP_RelTol
+    DT   = GET_TS_CHEM()          ! Chemistry timestep [s]
+    ATOL = State_Chm%KPP_AbsTol   ! Absolute tolerance
+    RTOL = State_Chm%KPP_RelTol   ! Relative tolerance
 
     !=======================================================================
     ! %%%%% SOLVE CHEMISTRY -- This is the main KPP solver loop %%%%%
@@ -681,6 +664,9 @@ CONTAINS
                 !    photolysis species index (range: 1..State_Chm%nPhotol)
                 !    for each of the FAST-JX photolysis species (range;
                 !    1..State_Chm%Phot%nMaxPhotRxns) in the GC_PHOTO_ID array
+                !
+                ! TODO: Abstract some of this to a subroutine,
+                !       to simplify DO_FULLCHEM
                 !============================================================
 
                 ! GC photolysis species index
@@ -770,6 +756,8 @@ CONTAINS
        ! This process has to be done before InChemGrid as it is supposed to
        ! be active everywhere, especially the stratosphere.
        ! (hplin, 5/30/23)
+       !
+       ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
        !=====================================================================
 
        IF ( Input_Opt%correctConvUTLS .and. L .ge. State_Met%PBL_TOP_L(I,J) ) THEN
@@ -962,7 +950,7 @@ CONTAINS
        !=====================================================================
 
        ! Update the array of rate constants
-       CALL Update_RCONST( )
+       CALL Update_RCONST()
 
        !=====================================================================
        ! HISTORY (aka netCDF diagnostics)
@@ -973,6 +961,8 @@ CONTAINS
        ! NOTE: In KPP 2.5.0+, VAR and FIX are now private to the integrator
        ! and point to C.  Therefore, pass C(1:NVAR) instead of VAR and
        ! C(NVAR+1:NSPEC) instead of FIX to routine FUN.
+       !
+       ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
        !=====================================================================
        IF ( State_Diag%Archive_RxnRate                                  .or. &
             State_Diag%Archive_SatDiagnRxnRate                        ) THEN
@@ -1028,7 +1018,6 @@ CONTAINS
        ! Set options for the LSODE integrator in vectors ICNTRL and RCNTRL
        ! This now needs to be done within the parallel loop
        !=====================================================================
-       ICNTRL(2)  = 5000                       ! Max number of steps
        ICNTRL(4)  = 2                          ! Order of integration
        ICNTRL(15) = -1                         ! No rate updates in integrator
        RCNTRL(3)  = State_Chm%KPPHvalue(I,J,L) ! Archived Hstart from restart
@@ -1048,7 +1037,9 @@ CONTAINS
        local_RCONST          = RCONST
 
        ! Call the KPP integrator
-       CALL Integrate( TIN, TOUT, ICNTRL, RCNTRL, ISTATUS, RSTATE, IERR )
+       ! NOTE: Some integrators (like LSODE) will overwrite the TIN value
+       ! upon exit.  To prevent this, pass 0.0, DT as the 1st 2 arguments.
+       CALL Integrate( 0.0_dp, DT, ICNTRL, RCNTRL, ISTATUS, RSTATE, IERR )
 
        ! Print grid box indices to screen if integrate failed
        IF ( IERR < 0 ) THEN
@@ -1092,7 +1083,7 @@ CONTAINS
        !=====================================================================
        ! HISTORY: Archive KPP solver diagnostics
        !
-       ! !TODO: Abstract this into a separate routine
+       ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
        !=====================================================================
        IF ( State_Diag%Archive_KppDiags ) THEN
 
@@ -1155,8 +1146,8 @@ CONTAINS
        !=====================================================================
        IF ( IERR < 0 ) THEN
 
-          ! Zero the first time step (Hstart, used by Rosenbrock).  Also reset
-          ! C with concentrations prior to the 1st call to "Integrate".
+          ! Zero the first time step (Hstart).  Also reset C with
+          ! concentrations prior to the 1st call to "Integrate".
           RCNTRL(3) = 0.0_dp
           C         = C_before_integrate
 
@@ -1172,11 +1163,12 @@ CONTAINS
           ENDIF
 
           ! Update rates again
-          CALL Update_RCONST( )
+          CALL Update_RCONST()
 
-          ! Call the Rosenbrock integrator (w/ auto-reduction disabled)
-          CALL Integrate( TIN,    TOUT,    ICNTRL,                           &
-                          RCNTRL, ISTATUS, RSTATE, IERR                     )
+          ! Call the integrator
+          ! NOTE: Some integrators (like LSODE) will overwrite the TIN value
+          ! upon exit.  To prevent this, pass 0.0, DT as the 1st 2 arguments.
+          CALL Integrate( 0.0_dp, DT, ICNTRL, RCNTRL, ISTATUS, RSTATE, IERR )
 
           !==================================================================
           ! HISTORY: Archive KPP solver diagnostics
@@ -1373,6 +1365,8 @@ CONTAINS
        ! Obtain P/L with a unit [kg S] for tracing
        ! gas-phase sulfur species production (SO2, SO4, MSA)
        ! (win, 8/4/09)
+       !
+       ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
        !-----------------------------------------------------------------
 
        ! Calculate H2SO4 production rate [kg s-1] in each
@@ -1418,6 +1412,8 @@ CONTAINS
 #ifdef MODEL_GEOS
        !--------------------------------------------------------------------
        ! Archive NOx lifetime [h]
+       !
+       ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
        !--------------------------------------------------------------------
        IF ( State_Diag%Archive_NoxTau .OR. State_Diag%Archive_TropNOxTau ) THEN
           CALL Fun( V       = C(1:NVAR),                                     &
@@ -1468,6 +1464,8 @@ CONTAINS
        ! NOTE: KppId is the KPP ID # for each of the prod and loss
        ! diagnostic species.  This is the value used to index the
        ! KPP "C" array (in module gckpp_Global.F90).
+       !
+       ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
        !====================================================================
 
        ! Chemical loss of species or families [molec/cm3/s]
@@ -1505,6 +1503,8 @@ CONTAINS
        !--------------------------------------------------------------------
        ! Archive prod/loss fields for the TagCO simulation [molec/cm3/s]
        ! (In practice, we only need to do this from benchmark simulations)
+       !
+       ! TODO: Abstract this to a subroutine, to simplify DO_FULLCHEM
        !--------------------------------------------------------------------
        IF ( State_Diag%Archive_ProdCOfromCH4     .or.                        &
             State_Diag%Archive_ProdCOfromNMVOC ) THEN